Move float functions to datatypes package
